Rename all projects files and folders
