--- src/Zedouto.Api.Facade/Zedouto.Api.Facade.csproj
@@ -1,17 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="System.IdentityModel.Tokens.Jwt" Version="6.8.0" />
-    <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="3.1.8" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\Zedouto.Api.Model\Zedouto.Api.Model.csproj" />
-    <ProjectReference Include="..\Zedouto.Api.Service\Zedouto.Api.Service.csproj" />
-  </ItemGroup>
-
-</Project>
