+++ src/Zedouto.Api.Login.Facade/Interfaces/IUserFacade.cs
@@ -1,0 +1,12 @@
+using System.Threading.Tasks;
+using Zedouto.Api.Login.Model.Entities;
+using Zedouto.Api.Login.Model.Models;
+
+namespace Zedouto.Api.Login.Facade.Interfaces
+{
+    public interface IUserFacade
+    {
+        Task AddUserAsync(User user);
+        Task<UserToken> LoginAsync(User user);
+    }
+}
