+++ src/Zedouto.Api.Login.Model/Configurations/ApplicationSettings.cs
@@ -1,0 +1,13 @@
+using Newtonsoft.Json;
+
+namespace Zedouto.Api.Login.Model.Configurations
+{
+    public class ApplicationSettings
+    {
+        [JsonProperty]
+        public string Name { get; set; }
+
+        [JsonProperty]
+        public string Version { get; set; }
+    }
+}
