+++ src/Zedouto.Api.Login.Model/Configurations/FirestoreAppSettings.cs
@@ -1,0 +1,16 @@
+using Newtonsoft.Json;
+
+namespace Zedouto.Api.Login.Model.Configurations
+{
+    public class FirestoreAppSettings
+    {
+        [JsonProperty]
+        public string ProjectId { get; set; }
+
+        [JsonProperty]
+        public string CollectionName { get; set; }
+
+        [JsonProperty]
+        public string Credentials { get; set; }
+    }
+}
