+++ src/Zedouto.Api.Login.Model/Entities/User.cs
@@ -1,0 +1,14 @@
+using Google.Cloud.Firestore;
+
+namespace Zedouto.Api.Login.Model.Entities
+{
+    [FirestoreData]
+    public class User
+    {
+        [FirestoreProperty]
+        public string Login { get; set; }
+
+        [FirestoreProperty]
+        public string Password { get; set; }
+    }
+}
