+++ src/Zedouto.Api.Login.Repository/Zedouto.Api.Login.Repository.csproj
@@ -1,0 +1,16 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Google.Cloud.Firestore" Version="2.2.0" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\Zedouto.Api.Login.Model\Zedouto.Api.Login.Model.csproj" />
+  </ItemGroup>
+
+
+</Project>
