+++ src/Zedouto.Api.Login.Service/Interfaces/IUserService.cs
@@ -1,0 +1,11 @@
+using System.Threading.Tasks;
+using Zedouto.Api.Login.Model.Entities;
+
+namespace Zedouto.Api.Login.Service.Interfaces
+{
+    public interface IUserService
+    {
+        Task AddUserAsync(User user);
+        Task<User> GetUserAsync(User user);
+    }
+}
