+++ src/Zedouto.Api.Login/Zedouto.Api.Login.csproj
@@ -1,8 +1,8 @@
 <Project Sdk="Microsoft.NET.Sdk.Web">
 
   <ItemGroup>
-    <ProjectReference Include="..\Zedouto.Api.Model\Zedouto.Api.Model.csproj" />
-    <ProjectReference Include="..\Zedouto.Api.Facade\Zedouto.Api.Facade.csproj" />
+    <ProjectReference Include="..\Zedouto.Api.Login.Model\Zedouto.Api.Login.Model.csproj" />
+    <ProjectReference Include="..\Zedouto.Api.Login.Facade\Zedouto.Api.Login.Facade.csproj" />
   </ItemGroup>
 
   <ItemGroup>
