--- src/Zedouto.Api.Model/Configurations/ApplicationSettings.cs
@@ -1,13 +1,0 @@
-using Newtonsoft.Json;
-
-namespace Zedouto.Api.Model.Configurations
-{
-    public class ApplicationSettings
-    {
-        [JsonProperty]
-        public string Name { get; set; }
-
-        [JsonProperty]
-        public string Version { get; set; }
-    }
-}
