--- src/Zedouto.Api.Model/Configurations/FirestoreAppSettings.cs
@@ -1,16 +1,0 @@
-using Newtonsoft.Json;
-
-namespace Zedouto.Api.Model.Configurations
-{
-    public class FirestoreAppSettings
-    {
-        [JsonProperty]
-        public string ProjectId { get; set; }
-
-        [JsonProperty]
-        public string CollectionName { get; set; }
-
-        [JsonProperty]
-        public string Credentials { get; set; }
-    }
-}
