--- src/Zedouto.Api.Model/Configurations/JwtConfigurationSettings.cs
@@ -1,13 +1,0 @@
-using Newtonsoft.Json;
-
-namespace Zedouto.Api.Model.Configurations
-{
-    public class JwtConfigurationSettings
-    {
-        [JsonProperty]
-        public string Key { get; set; }
-
-        [JsonProperty]
-        public int HourExpiration { get; set; }
-    }
-}
