--- src/Zedouto.Api.Model/Entities/User.cs
@@ -1,14 +1,0 @@
-using Google.Cloud.Firestore;
-
-namespace Zedouto.Api.Model.Entities
-{
-    [FirestoreData]
-    public class User
-    {
-        [FirestoreProperty]
-        public string Login { get; set; }
-
-        [FirestoreProperty]
-        public string Password { get; set; }
-    }
-}
