--- src/Zedouto.Api.Model/Models/UserToken.cs
@@ -1,10 +1,0 @@
-using System;
-
-namespace Zedouto.Api.Model.Models
-{
-    public class UserToken
-    {
-        public string Token { get; set; }
-        public DateTime Expiration { get; set; }
-    }
-}
