--- src/Zedouto.Api.Repository/Zedouto.Api.Repository.csproj
@@ -1,16 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Google.Cloud.Firestore" Version="2.2.0" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\Zedouto.Api.Model\Zedouto.Api.Model.csproj" />
-  </ItemGroup>
-
-
-</Project>
