--- src/Zedouto.Api.Service/Interfaces/ICryptographyService.cs
@@ -1,10 +1,0 @@
-using System.Threading.Tasks;
-
-namespace Zedouto.Api.Service.Interfaces
-{
-    public interface ICryptographyService
-    {
-        Task<string> CryptographAsync(string text);
-        Task<bool> EqualsTextAsync(string text, string cryptoText);
-    }
-}
