--- src/Zedouto.Api.Service/Interfaces/IUserService.cs
@@ -1,11 +1,0 @@
-using System.Threading.Tasks;
-using Zedouto.Api.Model.Entities;
-
-namespace Zedouto.Api.Service.Interfaces
-{
-    public interface IUserService
-    {
-        Task AddUserAsync(User user);
-        Task<User> GetUserAsync(User user);
-    }
-}
